Переносит код в файлы
